# More efficient parsing of UUIDs from strings