## Before
#### Corretto 8
```
[info] Benchmark                  Mode  Cnt       Score      Error  Units
[info] UUIDBenchmark.fromString  thrpt   15  948839.022 ± 7011.880  ops/s
```
#### Corretto 11
```
[info] Benchmark                  Mode  Cnt        Score       Error  Units
[info] UUIDBenchmark.fromString  thrpt   15  1500465.570 ± 11592.723  ops/s
```
#### OpenJDK 15
```
[info] Benchmark                  Mode  Cnt        Score       Error  Units
[info] UUIDBenchmark.fromString  thrpt   15  1826314.320 ± 33815.069  ops/s
```
## After
#### Corretto 8
```
[info] Benchmark                  Mode  Cnt         Score       Error  Units
[info] UUIDBenchmark.fromString  thrpt   15  21568831.647 ± 83625.539  ops/s
```
#### Corretto 11
```
[info] Benchmark                  Mode  Cnt         Score       Error  Units
[info] UUIDBenchmark.fromString  thrpt   15  23971815.448 ± 35417.588  ops/s
```
#### OpenJDK 15
```
[info] Benchmark                  Mode  Cnt         Score       Error  Units
[info] UUIDBenchmark.fromString  thrpt   15  27989191.275 ± 37117.812  ops/s
```